Added distinction between +50% and +99% solutions
Futae Yamagawa - https://steamcommunity.com/app/792100/discussions/0/1739968490573286109/#c1739968490575204652

During the beta, testers were met with moving targets for smallest/fastest solutions. Commands would be adjusted and suddenly work differently, or puzzles might change subtly (seek and destroy 3 changed 2x)

One of those targets that moved was the failure % for solutions. The early game allowed a level to clear if your solution worked at least half the time. It also had OCD scores that were in some cases easily met. Then mid beta reported OCD+ scores, but it excluded solutions that did not work 99%+, so Montecarlo solutions and some random walk solutions were omitted from the OCD+ scores. Later beta made it clear that this was happening by stating ‘>99% success’. I can’t speak for other testers, but seeing that, I put effort into making my shortest, fastest solutions all work 100% so that they would be included in the scoring. With the release, that restriction was relaxed to 50% - alas, I’d been making OCD solutions for the wrong goal, but it means players have more room to surpass the’best Known’.

While there were other changing targets in the beta, this one I think is relevant.

In the programming world, a solution that works only 50% won’t make it past QA, and would be abandoned in minutes by users. A solution that works on;y 99% may pass QA, but be abandoned by users in days. Even one failure in a million can be too many failures.

With that in mind, will there be a place for the fastest/shortest solutions that work 99+%, even if there are other, shorter or faster solutions that work 50+%?