Year 53 - 100 Cubes on the Floor (9 commands, 100%)
Just messed around with this a bit. It's actually quite interesting how it works.
Every worker except of the top one actually need to stand on the first cube of the row one cycle to set therir mem value correctly, just so that we can have 1 if less in there (which is why the and topright is not a worker is needed in the first if, otherwise there's one worker who does not stop one cycle on their cube).

So that's another one down to the "shortest known solution at release time" for the 99+ category.